move DetailTransactionActivity to package detail
